Set up initial connection stuff
Proof-of-concept confirmed; The webdav dependency works with Android out
of the box, and it works as intended. This does require disabling CORS
in copyparty first.
A quick kagi[^1] search found that some other DAV servers may not have
CORS enabled at all in the first place, so copyparty may be a one-off in
that regard, but it does not match well with discord.

That said, it now successfully lists the content of my WebDAV share :D

I also noticed the config is shared between devices, which is fine for
full vault sync imo. Pretty neat though :3

Client certificates (per my security model) are still a wildcard. I half
wonder if they'd work out of the box on android, but are automagically
fucked over by desktop operating systems due to variable keystores. Will
have to find out later, or find an alternate security model.
Work IPs should be fairly stable, so could potentially whitelist those
in nginx.

I _really_ want client certificates to work though; it would be the
single best option

[^1]: https://kagi.com