Refactor sync into a modular function, fix mtime (again)
Refactor: the sync function has a lot of commonalities. The only thing
that needs to be specified is:
* How to delete
* How to add a file

And you're good. This refactor is also necessary to enable per-folder
sync, as the exact same sync algorithm is required for folder sync. The
only difference is that it's done with a different set of local files.

This should also make it easier to write sync algorithm tests. That
said, I'm not particularly happy with the tests as they stand. #2 should
solve this

Mtime: I was right in a hunch; the mtime isn't synced. I suspect
combined testing with uploads and downloads obscured this bug the first
time around, but it's glaringly obvious with the tests I did today.

`X-OC-MTime` is a thing that works though, so at least with servers that
support it, it should be solved.
Would still be nice to have more fallback algorithms, but there's a
finite limit to how much arbitrary shit I can support. Hashing is still
unviable (per #1), and without deeper integrations with the editor to
control what's saved and changed in real-time (which is really fucking
annoying to do), it'll kill performance. This is good enough for now,
and I can reconsider if this ends up being popular enough that it's a
problem.
Support from the webdav package is limited too, and I do not feel like
shoehorning in stuff into the DAV request. I know there's a thing that
can be set in PROPFIND or whatever, but as far as I can tell, getting
that set up with the webdav package is going to be a Whole Thing:tm: